Add failure log reporter
Add createFailureLogReporter() to @logtape/testing so tests can
buffer LogTape records and report them only after failing callbacks by
default.  The reporter supports configurable report modes, custom sinks,
and scoped configuration without mutating global logger routing.

Split the testing package into recorder and reporter entry points and
publish @logtape/testing/recorder and @logtape/testing/reporter subpath
exports for both ESM and CommonJS consumers.

Update the testing manual, package README, Twoslash configuration, and
changelog entry for the PR #189 API surface.

Fixes https://github.com/dahlia/logtape/issues/186
https://github.com/dahlia/logtape/pull/189
