+++ packages/testing/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: "src/mod.ts",
+  entry: ["src/mod.ts", "src/recorder.ts", "src/reporter.ts"],
   dts: {
     sourcemap: true,
   },
